Move metric and collector matcher to pkg/monitoring/metrics/testing
