+++ pkg/monitoring/metrics/testing/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "collector_result_matcher.go",
+        "metric_matcher.go",
+    ],
+    importpath = "kubevirt.io/kubevirt/pkg/monitoring/metrics/testing",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//vendor/github.com/machadovilaca/operator-observability/pkg/operatormetrics:go_default_library",
+        "//vendor/github.com/onsi/gomega/format:go_default_library",
+        "//vendor/github.com/onsi/gomega/types:go_default_library",
+    ],
+)
