+++ tests/libmonitoring/BUILD.bazel
@@ -3,23 +3,21 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "metric_matcher.go",
         "prometheus.go",
         "scaling.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libmonitoring",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/monitoring/metrics/testing:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/clientcmd:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/framework/checks:go_default_library",
         "//tests/testsuite:go_default_library",
-        "//vendor/github.com/machadovilaca/operator-observability/pkg/operatormetrics:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/github.com/onsi/gomega/format:go_default_library",
         "//vendor/github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1:go_default_library",
         "//vendor/github.com/prometheus/client_golang/api/prometheus/v1:go_default_library",
         "//vendor/k8s.io/api/autoscaling/v1:go_default_library",
